Improves global search/replace workflow
Instead of using `:Gsearch` directly, which is a different search than ack,
these mapping support the workflow of searching via ack and then quickly
opening the quickfix list for replacement if replacement is desired.

This means that one could also use `:AckAdd` to construct a quickfix list
of multiple searches before initiating replacement.