Fixes ack bindings getting ruined by result sorting
For some reason chaining the sort command in the handler interferes with the
key mappings that ack.vim adds to the quickfix list. Running the sort as an
autocmd appears to not do that. In the most general case we probably don't
want the results to *always* get sorted (for instance if the output was
errors from a :make), but current workflows involving the quickfix list are
all search oriented, so this should be fine for now.