Adds supertab plugin
This allows tab to be used for completion during insert mode, without entirely
taking the key over.

Note: This uses <C-P> for completion, meaning that the cursor starts at the
bottom of the list, where the best match is. This is the opposite behavior of
the <C-N> I have been using to date for the completion, which probably explains
why I have had to scroll so far down the list so often.